pyproject.toml
[project]
name = "neural_search"
version = "0.2.0"
description = "Hybrid semantic search — BM25 + Dense + Web (Tavily) + Reranking + RRF"
readme = "README.md"
authors = [
    { name = "Aryanshr", email = "Arynshr@users.noreply.github.com" }
]
requires-python = ">=3.11"

dependencies = [
    # Document parsing
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",

    # Chunking
    "langchain-text-splitters>=0.2.0",

    # Token counting (replaces word-split hack)
    "tiktoken>=0.7.0",

    # Sparse retrieval
    "bm25s>=0.1.10",
    "nltk>=3.8.0",

    # Dense retrieval + reranking
    "sentence-transformers>=3.0.0",
    "qdrant-client>=1.9.0",

    # Web retrieval
    "tavily-python>=0.3.0",

    # Inference
    "groq>=0.9.0",

    # API
    "fastapi>=0.111.0",
    "uvicorn>=0.29.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # UI
    "streamlit>=1.35.0",

    # Utils
    "loguru>=0.7.2",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
]

[project.scripts]
neural-search-api = "neural_search.api.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/neural_search"]
